pyproject.toml
[project]
name = "statem"
version = "0.1.0"
description = "AI 剧情游戏后端"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.29.0",
]
